semi-functions-path-definition created, fixed the error faced in createJourneyId
