+++ creditcards/semi/constant.js
@@ -2,6 +2,8 @@
 
 const JOURNEY_NAME = 'SMART_EMI_JOURNEY';
 
+const CHANNEL = 'ADOBE_WEB';
+
 const ERROR_MSG = {
   mobileError: 'Enter valid mobile number',
 };
@@ -15,10 +17,14 @@ const OTP_TIMER = 30;
 
 const MAX_OTP_RESEND_COUNT = 3;
 
+const CURRENT_FORM_CONTEXT = {};
+
 export {
   JOURNEY_NAME,
   ERROR_MSG,
   OTP_TIMER,
   SEMI_ENDPOINTS,
   MAX_OTP_RESEND_COUNT,
+  CURRENT_FORM_CONTEXT,
+  CHANNEL,
 };
